# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk manajemen todo list dengan JWT authentication\nSetiap user punya todo list sendiri (terisolasi by user_id)",
        "title": "Todo API",
        "contact": {
            "name": "Jason",
            "url": "https://github.com/Tarquished"
        },
        "version": "1.0"
    },
    "host": "todo-api-production-74d1.up.railway.app",
    "basePath": "/",
    "paths": {
        "/hapus-todo": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Untuk menghapus suatu To-Do dengan ID yang diberikan\nDiverifikasi melalui JWT Token untuk mengecek kepemilikan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Menghapus To-Do berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessHandlerHapusUpdateTodo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Melakukan login dengan username dan password yang sudah ada dalam database\nPassword akan diverify dan akan diberikan JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Username dan password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.InputAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponLogin"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Mendaftarkan user baru dengan username dan password\nPassword akan di-hash menggunakan bcrypt sebelum disimpan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register user baru",
                "parameters": [
                    {
                        "description": "Username dan password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.InputAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ResponPesan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    }
                }
            }
        },
        "/tambah-todo": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendaftarkan To-Do ke dalam database dengan melakukan verify JWT terlebih dahulu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Membuat To-Do secara single",
                "parameters": [
                    {
                        "description": "Judul dan Prioritas",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.listTodo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessHandlerTodoSingle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    }
                }
            }
        },
        "/tambah-todo-batch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendaftarkan To-Do batch ke dalam database dengan melakukan verify JWT terlebih dahulu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Membuat To-Do secara batch",
                "parameters": [
                    {
                        "description": "List todo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.listTodo"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.listTodoBatch"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    }
                }
            }
        },
        "/todos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menampilkan isi dari semua To-Do yang dimiliki oleh user\nMemverifikasi ID dan token yang dipakai melalui JWT verification",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Menampilkan semua To-Do User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nomor halaman (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah data per halaman (default: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PageData"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    }
                }
            }
        },
        "/update-todo": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Untuk mengubah suatu To-Do dengan ID yang diberikan\nDiverifikasi melalui JWT Token untuk mengecek kepemilikan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "Mengupdate To-Do berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "listTodo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.listTodo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessHandlerHapusUpdateTodo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/main.ResponError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.InputAuth": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "main.PageData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.getTodo"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "main.ResponError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "main.ResponPesan": {
            "type": "object",
            "properties": {
                "pesan": {
                    "type": "string",
                    "example": "Berhasil menambahkan username ke database"
                }
            }
        },
        "main.SuccessHandlerHapusUpdateTodo": {
            "type": "object",
            "properties": {
                "pesan": {
                    "type": "string"
                }
            }
        },
        "main.SuccessHandlerTodoSingle": {
            "type": "object",
            "properties": {
                "judul": {
                    "type": "string"
                },
                "pesan": {
                    "type": "string"
                },
                "prioritas": {
                    "type": "string"
                }
            }
        },
        "main.SuccessResponLogin": {
            "type": "object",
            "properties": {
                "pesan": {
                    "type": "string",
                    "example": "Berhasil login!"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "main.getTodo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "judul": {
                    "type": "string"
                },
                "prioritas": {
                    "type": "string"
                }
            }
        },
        "main.listTodo": {
            "type": "object",
            "properties": {
                "judul": {
                    "type": "string"
                },
                "prioritas": {
                    "type": "string"
                }
            }
        },
        "main.listTodoBatch": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "judul": {
                    "type": "string"
                },
                "prioritas": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}